refactor: remove unused tagsError prop and enhance tag handling with new utility functions
