+++ src/components/Tags/utils/_TagsCommonFunctions.tsx
@@ -1,4 +1,8 @@
-import { PostView } from '@/types/Post';
+import { useRef, useState } from 'react';
+import { PostTag, PostView } from '@/types/Post';
+import { getUserProfile } from '@/services/userService';
+import { UserView } from '@/types/User';
+import { useDrawerClickOutside } from '@/hooks/useDrawerClickOutside';
 
 /**
  * Function to add a tag to a post
@@ -103,6 +107,178 @@ export const handleDeleteTag = async (
   setLoadingTags('');
 };
 
-export const TagsCommonFunctions = { handleAddTag, handleDeleteTag };
+/**
+ * Function to sanitize tag input
+ * @param value Input value
+ * @returns Sanitized value
+ */
+export const sanitizeTagInput = (value: string): string => {
+  return value.toLowerCase().replace(/\s/g, '').replace(/!/g, '');
+};
+
+/**
+ * Function to fetch profile images for tags
+ * @param tag Tag object
+ * @param pubky Current user pubky
+ * @returns Array of image URLs
+ */
+export const fetchProfileImages = async (tag: PostTag, pubky: string | null): Promise<string[]> => {
+  const images = await Promise.all(
+    tag.taggers.map(async (fromItem) => {
+      try {
+        const profile = await getUserProfile(fromItem, pubky ?? '');
+        return profile?.details?.image || '/images/webp/Userpic.webp';
+      } catch (error) {
+        return '/images/webp/Userpic.webp';
+      }
+    })
+  );
+  return images;
+};
+
+/**
+ * Function to update tags and timeline
+ * @param tag Tag to update
+ * @param isAdding Whether adding or removing tag
+ * @param tags Current tags
+ * @param pubky Current user pubky
+ * @param setTags Function to update tags
+ * @param setTimeline Function to update timeline
+ * @param post Current post
+ */
+export const updateTagsAndTimeline = (
+  tag: string,
+  isAdding: boolean,
+  tags: PostTag[],
+  pubky: string | null,
+  setTags: (tags: PostTag[]) => void,
+  setTimeline: (callback: (prev: PostView[]) => PostView[]) => void,
+  post: PostView
+) => {
+  let newTags = tags;
+  const existingTag = tags.find((t) => t.label === tag);
+  let uniqueTagsChange = 0;
+
+  if (isAdding) {
+    if (!existingTag) {
+      newTags = [
+        ...tags,
+        {
+          label: tag,
+          taggers_count: 1,
+          taggers: [pubky ?? ''],
+          relationship: true
+        }
+      ];
+      uniqueTagsChange = 1;
+    } else if (!existingTag.taggers.includes(pubky ?? '')) {
+      newTags = tags.map((t) =>
+        t.label === tag
+          ? {
+              ...t,
+              taggers_count: t.taggers_count + 1,
+              taggers: [...t.taggers, pubky ?? ''],
+              relationship: true
+            }
+          : t
+      );
+      if (existingTag.taggers_count === 0) uniqueTagsChange = 1;
+    }
+  } else {
+    newTags = tags.map((t) =>
+      t.label === tag
+        ? {
+            ...t,
+            taggers_count: t.taggers_count - 1,
+            taggers: t.taggers.filter((tg) => tg !== pubky),
+            relationship: false
+          }
+        : t
+    );
+    if (existingTag?.taggers_count === 1) uniqueTagsChange = -1;
+  }
+
+  setTags(newTags);
+  setTimeline((prev) =>
+    prev.map((p) =>
+      p.details.id === post.details.id
+        ? {
+            ...p,
+            tags: newTags,
+            counts: {
+              ...p.counts,
+              unique_tags: Math.max(0, p.counts.unique_tags + uniqueTagsChange)
+            }
+          }
+        : p
+    )
+  );
+};
+
+/**
+ * Hook to handle emoji picker
+ * @returns Emoji picker state and functions
+ */
+export const useEmojiPicker = () => {
+  const [showEmojis, setShowEmojis] = useState(false);
+  const wrapperRefEmojis = useRef<HTMLDivElement>(null);
+
+  useDrawerClickOutside(wrapperRefEmojis, () => setShowEmojis(false));
+
+  return {
+    showEmojis,
+    setShowEmojis,
+    wrapperRefEmojis
+  };
+};
+
+/**
+ * Function to fetch profiles for taggers
+ * @param taggers Array of tagger IDs
+ * @param pubky Current user pubky
+ * @param userProfiles Current user profiles
+ * @param setUserProfiles Function to update user profiles
+ * @param setFollowedUser Function to update followed users
+ * @param setInitLoadingFollowers Function to update loading state
+ */
+export const fetchProfiles = async (
+  taggers: string[],
+  pubky: string | null,
+  userProfiles: { [key: string]: UserView },
+  setUserProfiles: (callback: (prev: { [key: string]: UserView }) => { [key: string]: UserView }) => void,
+  setFollowedUser: (callback: (prev: { [key: string]: boolean }) => { [key: string]: boolean }) => void,
+  setInitLoadingFollowers: (value: boolean) => void
+) => {
+  setInitLoadingFollowers(true);
+  const profilesMap: { [key: string]: UserView } = {};
+  const followedMap: { [key: string]: boolean } = {};
+
+  await Promise.all(
+    taggers.map(async (userId) => {
+      if (userProfiles[userId]) return;
+      try {
+        const profile = await getUserProfile(userId, pubky ?? '');
+        profilesMap[userId] = profile;
+        followedMap[userId] = profile.relationship?.following ?? false;
+      } catch (error) {
+        console.error(`Error fetching profile for user ${userId}`, error);
+      }
+    })
+  );
+
+  setUserProfiles((prev) => ({ ...prev, ...profilesMap }));
+  setFollowedUser((prev) => ({ ...prev, ...followedMap }));
+  setInitLoadingFollowers(false);
+};
+
+export const TagsCommonFunctions = {
+  handleAddTag,
+  handleDeleteTag,
+  sanitizeTagInput,
+  fetchProfileImages,
+  updateTagsAndTimeline,
+  useEmojiPicker,
+  fetchProfiles
+};
 
 export default TagsCommonFunctions;
